Easy track and lapping system somewhat done
An unfinished form of the easy track and lapping system have been added. The easy track currently consists of a boomerang-shaped main track with a small shortcut and the lapping system consists of a Game Manager, a lapping line, and an invisible one-way collider to ensure the player actually goes around the track.

These systems are unfinished and the opponent AIs are not started yet.